quarter_1/review/requirements.txt
certifi==2024.8.30
pycurl==7.45.3
python-dotenv==1.0.1
